pyproject.toml
[project]
name = "talent-manager"
version = "0.1.0"
description = "AI Talent Manager System"
authors = [
    {name = "korku karsah"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "requests (>=2.32.4,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "google-api-python-client (>=2.175.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "openai (>=1.93.0,<2.0.0)",
    "elevenlabs (>=2.6.0,<3.0.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "gtts (>=2.5.4,<3.0.0)",
    "pygame (>=2.6.1,<3.0.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "imageio (>=2.37.0,<3.0.0)",
    "imageio-ffmpeg (>=0.6.0,<0.7.0)",
    "moviepy @ git+https://github.com/Zulko/moviepy.git",
    "flower (>=2.0.1,<3.0.0)",
]

[tool.poetry]
packages = [{ include = "core" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"
black = "^25.1.0"
flake8 = "^7.3.0"
mypy = "^1.16.1"
pre-commit = "^4.2.0"
pytest-cov = "^6.2.1"

